feat: move filter controls to their own component
