+++ jobmon_gui/src/configs/Axios.ts
@@ -1,7 +1,6 @@
-
 export const jobmonAxiosConfig = {
-            headers: {
-                'Content-Type': 'application/json',
-                'Accept': 'application/json'
-            }
-        }
+    headers: {
+        'Content-Type': 'application/json',
+        'Accept': 'application/json'
+    }
+}
